Move storage classes to new FileStorage module
